requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1 
dj-database-url==1.0.0 
dj-email-url==1.0.5 
dj-rest-auth==2.2.5 
Django==4.1 
django-allauth==0.48.0 
django-cache-url==3.4.2 
django-cors-headers==3.13.0 
djangorestframework==3.13.1
drf-spectacular==0.21.2 
environs==9.5.0 
gunicorn==20.1.0 
idna==3.4 
inflection==0.5.1 
jsonschema==4.4.0 
marshmallow==4.17.0 
oauthlib==3.2.2 
psycopg2-binary==2.9.3 
pycparser==2.21 
PyJWT==2.6.0 
pyrsistent==0.19.1 
python-dotenv==0.20.0 
python3-openid==3.2.0 
pytz==2022.2.1 
PyYAML==6.0 
requests==2.28.1 
requests-oauthlib==1.3.1 
sqlparse==0.4.2 
uritemplate==4.1.1 
urllib3==1.26.12 
whitenoise==6.2.0